Extract hint assertions
